not done yet

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: hharit <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/22 10:36:45 by hharit            #+#    #+#             */
-/*   Updated: 2021/11/25 21:43:28 by hharit           ###   ########.fr       */
+/*   Updated: 2021/11/26 07:52:44 by hharit           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,11 +41,11 @@ void	ft_conversion(char c, va_list ptr, int *count)
 	if (c == 'c')
 		*count += ft_putchar(va_arg(ptr, int));
 	if (c == 'x')
-		ft_convert_hex_dec(va_arg(ptr, int), "0123456789abcdef", count);
+		ft_convert_hex_dec(va_arg(ptr, long), "0123456789abcdef", count);
 	if (c == 'X')
-		ft_convert_hex_dec(va_arg(ptr, int), "0123456789ABCDEF", count);
+		ft_convert_hex_dec(va_arg(ptr, long), "0123456789ABCDEF", count);
 	if (c == 'u')
-		ft_putnbr_u(va_arg(ptr, unsigned long), count);
+		ft_putnbr_u(va_arg(ptr, long), count);
 }
 
 int	ft_printf(const char *str, ...)
